+++ services/merchant-acquirer/src/test/java/com/processing/merchantacquirer/service/MerchantProviderTest.java
@@ -7,6 +7,7 @@ import com.processing.merchantacquirer.domain.entity.Merchant;
 import com.processing.merchantacquirer.domain.entity.Scenario;
 import com.processing.merchantacquirer.repository.MerchantRepository;
 
+import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
@@ -21,7 +22,7 @@ public class MerchantProviderTest {
   private MerchantProvider merchantProvider;
 
   private final Scenario scenario =
-      new Scenario(List.of("5411", "5499"), 100, 3000, "08:00", "23:00", 99);
+      new Scenario(List.of("5411", "5499"), BigDecimal.valueOf(100), BigDecimal.valueOf(3000), "08:00", "23:00", 99);
 
   @BeforeEach
   void setUp() {
@@ -55,7 +56,7 @@ public class MerchantProviderTest {
             "5411",
             "grocery",
             "ACQ003",
-            15L,
+            BigDecimal.valueOf(15),
             145000);
     List<Merchant> expectedMerchants = List.of(merchant);
 
@@ -78,7 +79,7 @@ public class MerchantProviderTest {
             "5411",
             "grocery",
             "ACQ003",
-            15L,
+            BigDecimal.valueOf(15),
             145000);
     Merchant merchant2 = new Merchant(
             "MERCH00000000007",
@@ -86,7 +87,7 @@ public class MerchantProviderTest {
             "5499",
             "grocery",
             "ACQ003",
-            15L,
+            BigDecimal.valueOf(15),
             145000);
     List<Merchant> expectedMerchants = List.of(merchant1, merchant2);
 
@@ -108,7 +109,7 @@ public class MerchantProviderTest {
             "5411",
             "grocery",
             "ACQ003",
-            15L,
+            BigDecimal.valueOf(15),
             145000);
     Merchant merchant2 = new Merchant(
             "MERCH00000000007",
@@ -116,7 +117,7 @@ public class MerchantProviderTest {
             "5499",
             "grocery",
             "ACQ003",
-            15L,
+            BigDecimal.valueOf(15),
             145000);
     List<Merchant> expectedMerchants = List.of(merchant1, merchant2);
